fix: move issuePage components to Pages file
